+++ test/Sentry.DiagnosticSource.IntegrationTests/TestDbBuilder.cs
@@ -1,20 +1,20 @@
+using System.Data;
 using System.Data.Common;
+using Microsoft.Data.SqlClient;
 using Microsoft.EntityFrameworkCore;
 
 public static class TestDbBuilder
 {
-    public static async Task CreateTable(DbConnection connection)
+    public static async Task CreateTable(SqlConnection connection)
     {
         using var dbContext = GetDbContext(connection);
         await dbContext.Database.EnsureCreatedAsync();
         using var command = connection.CreateCommand();
-        command.CommandText = "create table MyTable (Value int);";
+        command.CommandText = "create table MyTable (Value nvarchar(100));";
         await command.ExecuteNonQueryAsync();
     }
 
-    private static int intData;
-
-    private static TestDbContext GetDbContext(DbConnection connection)
+    private static TestDbContext GetDbContext(SqlConnection connection)
     {
         var builder = new DbContextOptionsBuilder<TestDbContext>();
         builder.UseSqlServer(connection);
@@ -23,39 +23,39 @@ public static class TestDbBuilder
         return dbContext;
     }
 
-    public static async Task AddEfData(DbConnection connection)
+    public static async Task AddEfData(SqlConnection connection)
     {
         using var dbContext = GetDbContext(connection);
-        dbContext.TestEntities.Add(new TestEntity { Property = "Value" });
+        dbContext.TestEntities.Add(new TestEntity { Property = "SHOULD NOT APPEAR IN PAYLOAD" });
         await dbContext.SaveChangesAsync();
     }
 
-    public static async Task GetEfData(DbConnection connection)
+    public static async Task GetEfData(SqlConnection connection)
     {
         using var dbContext = GetDbContext(connection);
         await dbContext.TestEntities.ToListAsync();
     }
 
-    public static async Task AddData(DbConnection connection)
+    public static async Task AddData(SqlConnection connection)
     {
         using var command = connection.CreateCommand();
-        var addData = intData;
-        intData++;
-        command.CommandText = $@"
+        command.Parameters.AddWithValue("value", "SHOULD NOT APPEAR IN PAYLOAD");
+        command.CommandText = @"
 insert into MyTable (Value)
-values ({addData});";
+values (@value);";
         await command.ExecuteNonQueryAsync();
     }
 
-    public static async Task GetData(DbConnection connection)
+    public static async Task GetData(SqlConnection connection)
     {
-        var values = new List<int>();
+        var values = new List<string>();
         using var command = connection.CreateCommand();
-        command.CommandText = "select Value from MyTable";
+        command.Parameters.AddWithValue("value", "SHOULD NOT APPEAR IN PAYLOAD");
+        command.CommandText = "select Value from MyTable where Value = @value";
         using var reader = await command.ExecuteReaderAsync();
         while (await reader.ReadAsync())
         {
-            values.Add(reader.GetInt32(0));
+            values.Add(reader.GetString(0));
         }
     }
 }