+++ test/Sentry.EntityFramework.Tests/SentryCommandInterceptorTests.cs
@@ -1,9 +1,6 @@
-using System.Data.Common;
-using System.Data.Entity.Infrastructure.Interception;
-using Effort.Provider;
-
 namespace Sentry.EntityFramework.Tests;
 
+[Collection("Sequential")]
 public class SentryCommandInterceptorTests
 {
     private class Fixture
@@ -35,7 +32,7 @@ public class SentryCommandInterceptorTests
             Query = "Expected query string"
         };
 
-        var interceptor = SentryDatabaseLogging.UseBreadcrumbs(_fixture.QueryLogger, initOnce: false);
+        var interceptor = SentryDatabaseLogging.UseBreadcrumbs(_fixture.QueryLogger, initOnce: false)!;
 
         var command = new EffortCommand
         {
@@ -43,7 +40,7 @@ public class SentryCommandInterceptorTests
         };
 
         interceptor.NonQueryExecuting(command, new DbCommandInterceptionContext<int>());
-        _fixture.QueryLogger.Received(1).Log(expected.Query, BreadcrumbLevel.Debug);
+        _fixture.QueryLogger.Received(1).Log(expected.Query);
     }
 
     [Fact]
@@ -54,14 +51,14 @@ public class SentryCommandInterceptorTests
             Query = "Expected query string"
         };
 
-        var interceptor = SentryDatabaseLogging.UseBreadcrumbs(_fixture.QueryLogger, initOnce: false);
+        var interceptor = SentryDatabaseLogging.UseBreadcrumbs(_fixture.QueryLogger, initOnce: false)!;
 
         var command = new EffortCommand
         {
             CommandText = expected.Query
         };
 
-        interceptor.NonQueryExecuting(command, new DbCommandInterceptionContext<int> { Exception = new Exception() });
+        interceptor.NonQueryExecuting(command, new DbCommandInterceptionContext<int> { Exception = new() });
         _fixture.QueryLogger.Received(1).Log(expected.Query, BreadcrumbLevel.Error);
     }
 
@@ -73,7 +70,7 @@ public class SentryCommandInterceptorTests
             Query = "Expected query string"
         };
 
-        var interceptor = SentryDatabaseLogging.UseBreadcrumbs(_fixture.QueryLogger, initOnce: false);
+        var interceptor = SentryDatabaseLogging.UseBreadcrumbs(_fixture.QueryLogger, initOnce: false)!;
 
         var command = new EffortCommand
         {
@@ -81,7 +78,7 @@ public class SentryCommandInterceptorTests
         };
 
         interceptor.ReaderExecuting(command, new DbCommandInterceptionContext<DbDataReader>());
-        _fixture.QueryLogger.Received(1).Log(expected.Query, BreadcrumbLevel.Debug);
+        _fixture.QueryLogger.Received(1).Log(expected.Query);
     }
 
     [Fact]
@@ -92,14 +89,14 @@ public class SentryCommandInterceptorTests
             Query = "Expected query string"
         };
 
-        var interceptor = SentryDatabaseLogging.UseBreadcrumbs(_fixture.QueryLogger, initOnce: false);
+        var interceptor = SentryDatabaseLogging.UseBreadcrumbs(_fixture.QueryLogger, initOnce: false)!;
 
         var command = new EffortCommand
         {
             CommandText = expected.Query
         };
 
-        interceptor.ReaderExecuting(command, new DbCommandInterceptionContext<DbDataReader> { Exception = new Exception() });
+        interceptor.ReaderExecuting(command, new DbCommandInterceptionContext<DbDataReader> { Exception = new() });
         _fixture.QueryLogger.Received(1).Log(expected.Query, BreadcrumbLevel.Error);
     }
 
@@ -111,7 +108,7 @@ public class SentryCommandInterceptorTests
             Query = "Expected query string"
         };
 
-        var interceptor = SentryDatabaseLogging.UseBreadcrumbs(_fixture.QueryLogger, initOnce: false);
+        var interceptor = SentryDatabaseLogging.UseBreadcrumbs(_fixture.QueryLogger, initOnce: false)!;
 
         var command = new EffortCommand
         {
@@ -119,7 +116,7 @@ public class SentryCommandInterceptorTests
         };
 
         interceptor.ScalarExecuting(command, new DbCommandInterceptionContext<object>());
-        _fixture.QueryLogger.Received(1).Log(expected.Query, BreadcrumbLevel.Debug);
+        _fixture.QueryLogger.Received(1).Log(expected.Query);
     }
 
     [Fact]
@@ -130,14 +127,14 @@ public class SentryCommandInterceptorTests
             Query = "Expected query string"
         };
 
-        var interceptor = SentryDatabaseLogging.UseBreadcrumbs(_fixture.QueryLogger, initOnce: false);
+        var interceptor = SentryDatabaseLogging.UseBreadcrumbs(_fixture.QueryLogger, initOnce: false)!;
 
         var command = new EffortCommand
         {
             CommandText = expected.Query
         };
 
-        interceptor.ScalarExecuting(command, new DbCommandInterceptionContext<object> { Exception = new Exception() });
+        interceptor.ScalarExecuting(command, new DbCommandInterceptionContext<object> { Exception = new() });
         _fixture.QueryLogger.Received(1).Log(expected.Query, BreadcrumbLevel.Error);
     }
 
@@ -148,6 +145,6 @@ public class SentryCommandInterceptorTests
 
         _ = _fixture.DbContext.TestTable.FirstOrDefault();
 
-        _fixture.QueryLogger.Received().Log(Arg.Any<string>(), BreadcrumbLevel.Debug);
+        _fixture.QueryLogger.Received().Log(Arg.Any<string>());
     }
 }