+++ test/Sentry.Serilog.Tests/AspNetCoreIntegrationTests.cs
@@ -1,5 +1,3 @@
-using Sentry.AspNetCore.Tests;
-
 namespace Sentry.Serilog.Tests;
 
 [Collection(nameof(SentrySdkCollection))]
@@ -16,7 +14,7 @@ public class AspNetCoreIntegrationTests : SerilogAspNetSentrySdkTestFixture
 
         Handlers = new[] { handler };
         Build();
-        _ = await HttpClient.GetAsync(handler.Path);
+        await HttpClient.GetAsync(handler.Path);
 
         Assert.Contains(Events, e => e.Logger == "Microsoft.AspNetCore.Diagnostics.ExceptionHandlerMiddleware");
         Assert.Collection(Events, @event => Assert.Collection(@event.SentryExceptions, x => Assert.False(x.Mechanism?.Handled)));