+++ test/Sentry.Tests/Internals/Http/CachingTransportTests.cs
@@ -1,3 +1,4 @@
+using System.IO.Abstractions.TestingHelpers;
 using System.Net;
 using System.Net.Http;
 using System.Net.Sockets;
@@ -10,6 +11,7 @@ namespace Sentry.Tests.Internals.Http;
 public class CachingTransportTests
 {
     private readonly TestOutputDiagnosticLogger _logger;
+    private readonly IFileSystem _fileSystem = new FakeFileSystem();
 
     public CachingTransportTests(ITestOutputHelper testOutputHelper)
     {
@@ -20,13 +22,14 @@ public class CachingTransportTests
     public async Task WithAttachment()
     {
         // Arrange
-        using var cacheDirectory = new TempDirectory();
+        using var cacheDirectory = new TempDirectory(_fileSystem);
         var options = new SentryOptions
         {
             Dsn = ValidDsn,
             DiagnosticLogger = _logger,
             Debug = true,
-            CacheDirectoryPath = cacheDirectory.Path
+            CacheDirectoryPath = cacheDirectory.Path,
+            FileSystem = _fileSystem
         };
 
         Exception exception = null;
@@ -71,13 +74,14 @@ public class CachingTransportTests
     public async Task WorksInBackground()
     {
         // Arrange
-        using var cacheDirectory = new TempDirectory();
+        using var cacheDirectory = new TempDirectory(_fileSystem);
         var options = new SentryOptions
         {
             Dsn = ValidDsn,
             DiagnosticLogger = _logger,
             Debug = true,
-            CacheDirectoryPath = cacheDirectory.Path
+            CacheDirectoryPath = cacheDirectory.Path,
+            FileSystem = _fileSystem
         };
 
         using var innerTransport = new FakeTransport();
@@ -103,13 +107,14 @@ public class CachingTransportTests
     public async Task ShouldNotLogOperationCanceledExceptionWhenIsCancellationRequested()
     {
         // Arrange
-        using var cacheDirectory = new TempDirectory();
+        using var cacheDirectory = new TempDirectory(_fileSystem);
 
         var options = new SentryOptions
         {
             Dsn = ValidDsn,
             DiagnosticLogger = _logger,
             CacheDirectoryPath = cacheDirectory.Path,
+            FileSystem = _fileSystem,
             Debug = true
         };
 
@@ -146,7 +151,7 @@ public class CachingTransportTests
     public async Task ShouldLogOperationCanceledExceptionWhenNotIsCancellationRequested()
     {
         // Arrange
-        using var cacheDirectory = new TempDirectory();
+        using var cacheDirectory = new TempDirectory(_fileSystem);
         var loggerCompletionSource = new TaskCompletionSource<object>();
 
         _logger
@@ -162,6 +167,7 @@ public class CachingTransportTests
             Dsn = ValidDsn,
             DiagnosticLogger = _logger,
             CacheDirectoryPath = cacheDirectory.Path,
+            FileSystem = _fileSystem,
             Debug = true
         };
 
@@ -191,13 +197,14 @@ public class CachingTransportTests
     public async Task EnvelopeReachesInnerTransport()
     {
         // Arrange
-        using var cacheDirectory = new TempDirectory();
+        using var cacheDirectory = new TempDirectory(_fileSystem);
         var options = new SentryOptions
         {
             Dsn = ValidDsn,
             DiagnosticLogger = _logger,
             Debug = true,
-            CacheDirectoryPath = cacheDirectory.Path
+            CacheDirectoryPath = cacheDirectory.Path,
+            FileSystem = _fileSystem
         };
 
         using var innerTransport = new FakeTransport();
@@ -217,13 +224,14 @@ public class CachingTransportTests
     public async Task MaintainsLimit()
     {
         // Arrange
-        using var cacheDirectory = new TempDirectory();
+        using var cacheDirectory = new TempDirectory(_fileSystem);
         var options = new SentryOptions
         {
             Dsn = ValidDsn,
             DiagnosticLogger = _logger,
             Debug = true,
             CacheDirectoryPath = cacheDirectory.Path,
+            FileSystem = _fileSystem,
             MaxCacheItems = 2
         };
 
@@ -245,13 +253,14 @@ public class CachingTransportTests
     public async Task AwareOfExistingFiles()
     {
         // Arrange
-        using var cacheDirectory = new TempDirectory();
+        using var cacheDirectory = new TempDirectory(_fileSystem);
         var options = new SentryOptions
         {
             Dsn = ValidDsn,
             DiagnosticLogger = _logger,
             Debug = true,
-            CacheDirectoryPath = cacheDirectory.Path
+            CacheDirectoryPath = cacheDirectory.Path,
+            FileSystem = _fileSystem
         };
 
         // Send some envelopes with a failing transport to make sure they all stay in cache
@@ -284,13 +293,14 @@ public class CachingTransportTests
     public async Task NonTransientExceptionShouldLog()
     {
         // Arrange
-        using var cacheDirectory = new TempDirectory();
+        using var cacheDirectory = new TempDirectory(_fileSystem);
         var options = new SentryOptions
         {
             Dsn = ValidDsn,
             DiagnosticLogger = _logger,
             Debug = true,
-            CacheDirectoryPath = cacheDirectory.Path
+            CacheDirectoryPath = cacheDirectory.Path,
+            FileSystem = _fileSystem
         };
 
         var innerTransport = Substitute.For<ITransport>();
@@ -320,13 +330,14 @@ public class CachingTransportTests
     public async Task DoesNotRetryOnNonTransientExceptions()
     {
         // Arrange
-        using var cacheDirectory = new TempDirectory();
+        using var cacheDirectory = new TempDirectory(_fileSystem);
         var options = new SentryOptions
         {
             Dsn = ValidDsn,
             DiagnosticLogger = _logger,
             Debug = true,
-            CacheDirectoryPath = cacheDirectory.Path
+            CacheDirectoryPath = cacheDirectory.Path,
+            FileSystem = _fileSystem
         };
 
         var innerTransport = Substitute.For<ITransport>();
@@ -365,13 +376,14 @@ public class CachingTransportTests
     public async Task RecordsDiscardedEventOnNonTransientExceptions()
     {
         // Arrange
-        using var cacheDirectory = new TempDirectory();
+        using var cacheDirectory = new TempDirectory(_fileSystem);
         var options = new SentryOptions
         {
             Dsn = ValidDsn,
             DiagnosticLogger = _logger,
             Debug = true,
             CacheDirectoryPath = cacheDirectory.Path,
+            FileSystem = _fileSystem,
             ClientReportRecorder = Substitute.For<IClientReportRecorder>()
         };
 
@@ -397,13 +409,14 @@ public class CachingTransportTests
     public async Task RoundtripsClientReports()
     {
         // Arrange
-        using var cacheDirectory = new TempDirectory();
+        using var cacheDirectory = new TempDirectory(_fileSystem);
         var options = new SentryOptions
         {
             Dsn = ValidDsn,
             DiagnosticLogger = _logger,
             Debug = true,
-            CacheDirectoryPath = cacheDirectory.Path
+            CacheDirectoryPath = cacheDirectory.Path,
+            FileSystem = _fileSystem
         };
 
         var timestamp = DateTimeOffset.UtcNow;
@@ -443,13 +456,14 @@ public class CachingTransportTests
     public async Task RestoresDiscardedEventCounts()
     {
         // Arrange
-        using var cacheDirectory = new TempDirectory();
+        using var cacheDirectory = new TempDirectory(_fileSystem);
         var options = new SentryOptions
         {
             Dsn = ValidDsn,
             DiagnosticLogger = _logger,
             Debug = true,
-            CacheDirectoryPath = cacheDirectory.Path
+            CacheDirectoryPath = cacheDirectory.Path,
+            FileSystem = _fileSystem
         };
 
         var recorder = (ClientReportRecorder) options.ClientReportRecorder;
@@ -496,13 +510,14 @@ public class CachingTransportTests
     public async Task TestNetworkException(Exception exception)
     {
         // Arrange
-        using var cacheDirectory = new TempDirectory();
+        using var cacheDirectory = new TempDirectory(_fileSystem);
         var options = new SentryOptions
         {
             Dsn = ValidDsn,
             DiagnosticLogger = _logger,
             Debug = true,
-            CacheDirectoryPath = cacheDirectory.Path
+            CacheDirectoryPath = cacheDirectory.Path,
+            FileSystem = _fileSystem
         };
 
         var receivedException = new Exception();
@@ -535,7 +550,7 @@ public class CachingTransportTests
 
         // Assert
         Assert.Equal(exception, receivedException);
-        Assert.True(Directory.EnumerateFiles(cacheDirectory.Path, "*", SearchOption.AllDirectories).Any());
+        Assert.True(_fileSystem.EnumerateFiles(cacheDirectory.Path, "*", SearchOption.AllDirectories).Any());
     }
 
     [Fact]
@@ -547,13 +562,14 @@ public class CachingTransportTests
         listener.WaitForNetworkOnlineAsync(Arg.Any<CancellationToken>())
             .Throws(new Exception("We should not be waiting for the network status if we know it's online."));
 
-        using var cacheDirectory = new TempDirectory();
+        using var cacheDirectory = new TempDirectory(_fileSystem);
         var options = new SentryOptions
         {
             Dsn = ValidDsn,
             DiagnosticLogger = _logger,
             Debug = true,
             CacheDirectoryPath = cacheDirectory.Path,
+            FileSystem = _fileSystem,
             NetworkStatusListener = listener
         };
 
@@ -584,13 +600,14 @@ public class CachingTransportTests
                 return Task.Delay(Timeout.Infinite, callInfo.Arg<CancellationToken>());
             });
 
-        using var cacheDirectory = new TempDirectory();
+        using var cacheDirectory = new TempDirectory(_fileSystem);
         var options = new SentryOptions
         {
             Dsn = ValidDsn,
             DiagnosticLogger = _logger,
             Debug = true,
             CacheDirectoryPath = cacheDirectory.Path,
+            FileSystem = _fileSystem,
             NetworkStatusListener = listener
         };
 
@@ -638,13 +655,14 @@ public class CachingTransportTests
                 return Task.CompletedTask;
             });
 
-        using var cacheDirectory = new TempDirectory();
+        using var cacheDirectory = new TempDirectory(_fileSystem);
         var options = new SentryOptions
         {
             Dsn = ValidDsn,
             DiagnosticLogger = _logger,
             Debug = true,
             CacheDirectoryPath = cacheDirectory.Path,
+            FileSystem = _fileSystem,
             NetworkStatusListener = listener
         };
 
@@ -659,5 +677,36 @@ public class CachingTransportTests
         // Assert
         var envelopes = innerTransport.GetSentEnvelopes();
         envelopes.Should().NotBeEmpty();
+    }
+
+    [Fact]
+    public async Task DoesntWriteSentAtHeaderToCacheFile()
+    {
+        // Arrange
+        using var cacheDirectory = new TempDirectory(_fileSystem);
+        var options = new SentryOptions
+        {
+            Dsn = ValidDsn,
+            DiagnosticLogger = _logger,
+            Debug = true,
+            CacheDirectoryPath = cacheDirectory.Path,
+            FileSystem = _fileSystem
+        };
+
+        var innerTransport = Substitute.For<ITransport>();
+        await using var transport = CachingTransport.Create(innerTransport, options, startWorker: false);
+
+        using var envelope = Envelope.FromEvent(new SentryEvent());
+
+        // Act
+        await transport.SendEnvelopeAsync(envelope);
+
+        // Assert
+        var filePath = _fileSystem
+            .EnumerateFiles(cacheDirectory.Path, "*", SearchOption.AllDirectories)
+            .Single();
+
+        var contents = _fileSystem.ReadAllTextFromFile(filePath);
+        Assert.DoesNotContain("sent_at", contents);
     }
 }