+++ test/Sentry.Tests/Protocol/SentryEventTests.cs
@@ -48,13 +48,18 @@ public class SentryEventTests
             SentryThreads = new[] { new SentryThread { Crashed = true } },
             ServerName = "server_name",
             TransactionName = "transaction",
-            DebugImages = new List<DebugImage>()
+            DebugImages = new List<DebugImage>
             {
-                new DebugImage { Type = "wasm", DebugId = "900f7d1b868432939de4457478f34720" }
+                new()
+                {
+                    Type = "wasm",
+                    DebugId = "900f7d1b868432939de4457478f34720"
+                }
             },
         };
 
         sut.Sdk.AddPackage(new Package("name", "version"));
+        sut.Sdk.AddIntegration("integration");
         sut.AddBreadcrumb(new Breadcrumb(timestamp, "crumb"));
         sut.AddBreadcrumb(new Breadcrumb(
             timestamp,